requirements.txt
# generated from manifests external_dependencies
invoice2data
odoorpc
